pyproject.toml
[project]
name = "adscope"
version = "0.1.0"
description = "한국 디지털 광고 통합 모니터링 인텔리전스 플랫폼"
requires-python = ">=3.12"
dependencies = [
    # 크롤러
    "playwright>=1.49",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    # 스케줄러
    "apscheduler>=3.10",
    # 백엔드
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pydantic>=2.10",
    "pydantic-settings>=2.6",
    "pyjwt>=2.9",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    # 데이터 처리
    "pandas>=2.2",
    "numpy>=2.1",
    "rapidfuzz>=3.10",
    # 캐싱
    "redis>=5.2",
    # 유틸
    "python-dotenv>=1.0",
    "loguru>=0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-playwright>=0.6",
    "black>=24.10",
    "isort>=5.13",
    "flake8>=7.1",
    "ruff>=0.8",
]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.backends._legacy:_Backend"
